Refactor middleware structure by moving multer and morgan middlewares to a common directory
