+++ src/middlewares/common/morgan.middleware.js
@@ -1,5 +1,5 @@
 import morgan from "morgan";
-import logger from "../utils/logger.js";
+import logger from "../../utils/logger.js";
 
 export const morganMiddleware = morgan(
   ":method :url :status :res[content-length] - :response-time ms",
