+++ src/routes/v1/deliveryPartner.routes.js
@@ -10,7 +10,7 @@ import {
   getDeliveryPartnerAccount,
 } from "../../controllers/v1/deliveryPartner.controller.js";
 import { verifyUserJwt } from "../../middlewares/auth.middleware.js";
-import { upload } from "../../middlewares/multer.middleware.js";
+import { upload } from "../../middlewares/common/multer.middleware.js";
 
 const router = Router();
